requirements.txt
apache-airflow[kubernetes]
google-api-python-client
google-cloud-storage